+++ app/src/main/java/fr/smarquis/qrcode/model/Mode.kt
@@ -1,3 +1,3 @@
-package fr.smarquis.qrcode
+package fr.smarquis.qrcode.model
 
 enum class Mode { AUTO, MANUAL }